pyproject.toml
[project]
name = "hf-summarizer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.42.59",
    "datasets>=4.6.1",
    "ensure==1.0.4",
    "evaluate>=0.4.6",
    "fastapi==0.78.0",
    "jinja2==3.1.2",
    "matplotlib>=3.10.8",
    "mypy-boto3-s3>=1.42.37",
    "nltk>=3.9.3",
    "notebook>=7.5.4",
    "pandas>=3.0.1",
    "py7zr>=1.1.0",
    "python-box==6.0.2",
    "pyyaml>=6.0.3",
    "rouge-score>=0.1.2",
    "sacrebleu>=2.6.0",
    "torch>=2.10.0",
    "tqdm>=4.67.3",
    "transformers[sentencepiece]>=5.2.0",
    "uvicorn==0.18.3",
]
